LayoutTests/platform/mac-wk1/fast/scrolling/scroll-animator-basic-events-expected.txt

CONSOLE MESSAGE: RenderLayer: mouseEnteredContentArea
CONSOLE MESSAGE: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: RenderLayer: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: RenderLayer: mouseExitedContentArea
PASS successfullyParsed is true

TEST COMPLETE
This is a scrollable div.
